adapter/requirements.txt
paho-mqtt==1.6.1
influxdb-client
python-dotenv
